fixing the product detail
